requirements.txt
selenium
openai
pytrends
requests
beautifulsoup4
feedparser
python-dotenv